refactor(render): extract a leaf package for config rendering
Rendering general.yaml and secrets.yaml lived in pkg/config/prompt, which
reaches 2218 transitive packages — ArgoCD, gitops-engine, casbin, a GitHub
App client — because PromptedConfig embedded config.ObmondoConfig and the
package sits alongside the interactive huh prompts. Anything wanting to
render a config had to take all of it, plus copies of this module's 17
replace directives, since Go honours replace only in the main module.

The Obmondo API renders the same two files server-side. Faced with that
import surface it vendored the templates instead, along with 14 golden
fixtures and a parity test whose only job is detecting drift between the
copy and this repo.

Move the render path into pkg/render:

  PromptedConfig, ObmondoConfig, RobotVSwitch, the four template methods,
  Render, renderTemplate, both templates, and the golden fixtures.

  go list -deps ./pkg/render  →  3 packages, all our own
                                 (render, constants, urlprotocol)
                                 75 including stdlib, zero third-party

Nothing external means no replace directive can apply, so importing this is
free for any module.

Two supporting moves. ObmondoConfig now lives in pkg/render because
PromptedConfig embeds it and leaving it in pkg/config (~1180 packages) would
have dragged that along; pkg/config aliases it, so its callers and yaml tags
are unchanged. The URL protocol classification moves to pkg/urlprotocol
because pkg/repository/url's other file reaches oklog/ulid, samber/lo and
golang.org/x/text, none of which a template needs to decide whether a fork
URL is SSH-form; pkg/repository/url re-exports it as aliases.

pkg/config/prompt re-exports PromptedConfig and Render, so the CLI's own
call sites are untouched. Type aliases, not defined types, because existing
code assigns and takes addresses across package boundaries.

writeConfigFiles stays with the prompt flow, and the assertion that it is a
thin caller of Render moves with it as TestWriteConfigFilesMatchesRender —
the golden fixtures pin Render, that pins the CLI's disk path against it.

No behaviour change: the golden fixtures are untouched and still pass.
