+++ pkg/render/templates/secrets.yaml.tmpl
