+++ pkg/render/testdata/golden/aws-workload.general.yaml
