+++ pkg/render/testdata/golden/aws-workload.secrets.yaml
