+++ pkg/render/testdata/golden/azure-workload.general.yaml
