+++ pkg/render/testdata/golden/azure-workload.secrets.yaml
