+++ pkg/render/testdata/golden/baremetal-generic-workload.general.yaml
