+++ pkg/render/testdata/golden/baremetal-generic-workload.secrets.yaml
