+++ pkg/render/testdata/golden/hetzner-baremetal-workload.general.yaml
