+++ pkg/render/testdata/golden/hetzner-baremetal-workload.secrets.yaml
