+++ pkg/render/testdata/golden/hetzner-hcloud-workload.general.yaml
