+++ pkg/render/testdata/golden/hetzner-hcloud-workload.secrets.yaml
