+++ pkg/render/testdata/golden/hetzner-hybrid-vpn.general.yaml
