+++ pkg/render/testdata/golden/hetzner-hybrid-vpn.secrets.yaml
