+++ pkg/render/testdata/golden/local-with-obmondo.general.yaml
