+++ pkg/render/testdata/golden/local-with-obmondo.secrets.yaml
